New database

+++ QLNhaHang/QLNhaHang/Models/Cartdetail.cs
@@ -8,7 +8,7 @@ namespace QLNhaHang.Models
         public int IdCart { get; set; }
         public int IdDish { get; set; }
         public int Quantity { get; set; }
-        public int Subtotal { get; set; }
+        public double Subtotal { get; set; }
 
         public virtual Cart IdCartNavigation { get; set; } = null!;
         public virtual Dish IdDishNavigation { get; set; } = null!;
