+++ QLNhaHang/QLNhaHang/Models/Workshift.cs
@@ -12,7 +12,7 @@ namespace QLNhaHang.Models
 
         public byte IdWorkshift { get; set; }
         public byte Shift { get; set; }
-        public byte Shifthour { get; set; }
+        public double Shifthour { get; set; }
 
         public virtual ICollection<Weeklyshift> Weeklyshifts { get; set; }
     }
